[SQLLIB][DOCS] Improve documentation for DECIMAL and FP; move FP functions to their own file
