+++ crates/sqllib/Cargo.toml
@@ -34,7 +34,7 @@ paste = { workspace = true }
 regex = { workspace = true }
 rkyv = { workspace = true, features = ["std", "size_64", "validation", "smallvec"] }
 hex = { workspace = true }
-num-traits = { workspace = true }
+num-traits = { workspace = true, features = ["std"] }
 flate2 = { workspace = true }
 metrics = { workspace = true }
 base64 = { workspace = true }
